+++ app/src/main/java/com/gomaping/navigation/ui/events/eventcard/TypeReview.kt
@@ -1,4 +1,4 @@
-package com.gomaping.navigation.ui.events.eventcar
+package com.gomaping.navigation.ui.events.eventcard
 
 enum class TypeReview(val type: Int) {
     DISGUSTING(1), BADLY(2), NOT_BAD(3), GOOD(4), EXCELLENT(5)
