+++ app/src/main/java/com/gomaping/navigation/ui/events/eventcard/adapter/CardImageAdapter.kt
@@ -1,4 +1,4 @@
-package com.gomaping.navigation.ui.events.eventcar.adapter
+package com.gomaping.navigation.ui.events.eventcard.adapter
 
 import android.view.LayoutInflater
 import android.view.View
@@ -6,19 +6,20 @@ import android.view.ViewGroup
 import android.widget.ImageView
 import androidx.recyclerview.widget.RecyclerView
 import com.gomaping.R
-import com.gomaping.navigation.ui.events.eventcar.ImageModel
+import com.gomaping.navigation.ui.events.eventcard.ImageModel
 
-class CardReviewAdapter(private val photo: List<ImageModel>) :
-    RecyclerView.Adapter<CardReviewAdapter.ImageHolder>() {
+class CardImageAdapter (private val photo: List<ImageModel>) :
+    RecyclerView.Adapter<CardImageAdapter.ImageHolder>() {
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ImageHolder {
         val view =
-            LayoutInflater.from(parent.context).inflate(R.layout.item_event_photo, parent, false)
+            LayoutInflater.from(parent.context).inflate(R.layout.card_image_view_pager, parent, false)
         return ImageHolder(view)
     }
 
     override fun onBindViewHolder(holder: ImageHolder, position: Int) {
-        holder.itemView.findViewById<ImageView>(R.id.image_review_card)
-            .setImageResource(photo[position].iconResId)
+        holder.itemView.findViewById<ImageView>(R.id.imageView).let {
+            it.setImageResource(photo[position].iconResId)
+        }
     }
 
     override fun getItemCount(): Int {
