+++ app/src/main/res/layout/fragment_reviews_event.xml
@@ -6,7 +6,7 @@
     android:layout_height="wrap_content"
     android:background="@android:color/transparent"
     android:orientation="vertical"
-    tools:context=".navigation.ui.events.eventcar.ScheduleEventFragment">
+    tools:context=".navigation.ui.events.eventcard.ScheduleEventFragment">
 
     <com.google.android.material.card.MaterialCardView
         android:layout_width="match_parent"
