+++ app/src/main/res/layout/item_event_review.xml
@@ -119,12 +119,11 @@
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/review_photo"
-            android:layout_width="match_parent"
+            android:layout_width="wrap_content"
             android:layout_height="70dp"
             android:layout_marginTop="14dp"
             android:orientation="horizontal"
             app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-            app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@id/review_event_text"
             tools:itemCount="3"
